Minor update to LockFreePQ.cpp test benchmark, prints longs instead of doubles for easier reading of results

+++ muse/kernel/src/mpi-mt-shm/LockFreePQ.cpp
@@ -594,7 +594,7 @@ void test(LockFreePQ *pq, std::vector<double> vals) {
         out += *val;
         delete  val;
     }
-    std::cout << "Inserted: " << in << " Got: " << out << " Difference: " << (in - out) << std::endl;
+    std::cout << "Inserted: " << (long)in << " Got: " << (long)out << " Difference: " << ((long)in - (long)out) << std::endl;
 }
 
 int main(int argc, char** argv) {
